requirements.txt
python-telegram-bot==13.0
requests
anthropic
flake8
black
pymongo[srv]
python-dotenv